Bake transforms from reference meshes, frames, and cubes onto final composed prim
Previously reference mesh transforms were maintained separately and composed via xformOpOrder, with a custom suffix. This causes issues in consuming applications, where manipulator tooling often only handles a single set of non-suffixed XformOps. The composed XformOp values are now baked into the composing prim's XformOps.

Unfortunately, this baking process requires special handling for UsdGeomCube, where a custom sizing is required to be computed separately from the position & orient XformOps... We need to make sure to treat this is the "newly applied transform" rather than multiply it as a "pre-exisitng local frame".

Finally, MjsFrame meta-elements are baked out when available. These are not considered part of the kinematic tree in MuJoCo and would not re-serialize in a straight MJCF->MjSpec->MJCF roundtrip, so we similarly bake out the values rather than introduce unexpected hierarchy.