Lock to mujoco 3.3.5
mujoco 3.3.6 introduced a breaking change to the compiler flags which is not reflected in the USD schema. we need to wait for a schema fix before moving up

Note also the schema itself is locked to 3.3.5 in install_mjc_schema so we really should match that here. Ideally that script could disappear if mujoco ships the schema in their own package at some point...